changes (#32)
This change is a temporary hack that we unfortunately have to do until we settle on a more sustainable, correct solution when we have more time to work on this.  The grpc-gateway plugin for protoc does not support the `source_relative` option that the normal golang plugin supports.  As such, the `gw.pb.go` files are being generated in the wrong location.  This moves them to the correct location, and deletes the folder structure.

We should contribute to that plugin and add that option when we have more time.

@kumare3 @anandswaminathan