Generate JS/TS modules (#43)
I created a docker image which embeds protobufjs and allow us to pass all the correct arguments to it. I'll be pushing that to a separate repo. But for now, I published a version of it to the internal docker registry and used it to generate the JS/TS modules in the same formats used by lyft-idl in flyte 1.0 .