Change Variable (#39)
The name field is redundant since a variable should always exist within the context of a map.  I also think the comment on it was out of date because variables cannot be linked by Variable proto alone.--there must be a link between nodes.

Renamed the fields for default to be less confusing.  Prior to this change, one would set 'value' to set a default value.