Create Email Message for publisher/subcriber (#94)
Notification information is split between common.proto and notification.proto for now - that will be consolidated after feature parity.

notification.proto contains messages that are used for the publisher/subscriber. EmailMessage will contain the relevant information for sending emails so the subscriber needs to unmarshall it and send it away. The publisher will populate the fields prior to publishing it.