Breaking: Improve Blob Typing (#140)
Both Flyte 1.0 and Flyte 2.0 SDKs enforce special behavior for blob objects that differ in format or dimensionality.  However, the typing information associated with a blob currently in 2.0 does not maintain enough information to enforce this differentiation.  This makes it problematic to implement since we need to wait until we receive an actual literal to bind to the correct implementation.  This PR treats a blob more along the likes of a schema in that there is meta-type information associated with and the implementation can decide how to handle the additional information.  So for instance, our implementation of Blob and MultiPartBlob expect dimensionalities as one might infer, but another implementation could treat all blob formats and dimensionalities as one object type that the user has to make decisions about how to interpret based on the associated type metadata.

This is a breaking change, but it should be extremely limited in scope of affected components since the new type information is additive and no one, except SDK and perhaps propeller currently, interprets down to this level.  cc: @schottra @katrogan @kumare3