Add Type Metadata (#156)
For SDK to meet it's extensibility goals, users need to be able to determine specialized behavior for a given type.  To give a practical example, imagine a user wants to create a  `Types.Proto(SomeProtobufObject)`.  Since we are system owners, we could just add a type to the spec like:

```
message ProtoType {
    string protobuf_package = 1;
    string protobuf_name = 2;
}
```

However, that is excessive for a user and a lot of users wouldn't have permission to alter the spec anyway.  By adding a metadata field, the actual type would merely be a binary, but the type engine can read the metadata field which would contain the information necessary to identify the protobuf.  In Python, this would be the protobuf package and name.  The type engine could then attempt to load the protobuf class and deserialize the binary.  An engine that doesn't want to specify behavior specific to protobufs, would receive it merely as a binary.

This is how I intend to implement protobuf types for SDK as a specialization of binary.