Distinguish Notification Overrides (#167)
We need to be able to distinguish between "this is an empty list of notification overrides so please use the default notifications" from "I want to override the default notifications to not create any notifications."

This is a breaking change in the sense that admin will need to update it's code upon upgrade to the new IDL.  However, it is not breaking of the data model.  Any existing notification overrides set in an execution spec will be lost to history.  However, the good news is that this wasn't plumbed through yet from SDK, CLI, UI, or anything so there is likely no data stored in the field we are removing.  So although it would wipe data, there is no data to wipe.