Add generation of swagger to code using go-bindata (#191)
This uses the go-bindata tool installed in the protocgenerator image to turn the generated `admin.swagger.json` file into a `.go` source code file that has the binary form of the json file embedded in it.  This file, generated in the same folder as the swagger file, is then serve-able by Flyte Admin.